Refactoring: moving DTO to controller package. DTOs are here for the view layer. They shouldn't impact the model. You could have more than one DTO for each domain class depending on the use case. It is the responsability of the controller to build these DTOs from the model.
