+++ src/main/kotlin/mixit/controller/GlobalController.kt
@@ -1,8 +1,7 @@
 package mixit.controller
 
-import mixit.model.Role
+import mixit.model.*
 import mixit.model.SponsorshipLevel.*
-import mixit.model.toDto
 import mixit.repository.EventRepository
 import mixit.repository.UserRepository
 import mixit.util.MarkdownConverter
@@ -14,6 +13,7 @@ import org.springframework.stereotype.Controller
 import org.springframework.web.reactive.function.server.Routes
 import org.springframework.web.reactive.function.server.ServerRequest
 import org.springframework.web.reactive.function.server.ServerResponse.ok
+import java.time.LocalDate
 import java.util.*
 
 
@@ -48,3 +48,43 @@ class GlobalController(val userRepository: UserRepository,
     }
 
 }
+
+fun EventSponsoring.toDto() = SponsorDto(
+        this.sponsor.login,
+        this.sponsor.company!!,
+        this.sponsor.logoUrl!!,
+        logoType(this.sponsor.logoUrl),
+        logoWebpUrl(this.sponsor.logoUrl)
+)
+
+private fun logoWebpUrl(url: String) =
+        when {
+            url.endsWith("png") -> url.replace("png", "webp")
+            url.endsWith("jpg") -> url.replace("jpg", "webp")
+            else -> null
+        }
+
+private fun logoType(url: String) =
+        when {
+            url.endsWith("svg") -> "image/svg+xml"
+            url.endsWith("png") -> "image/png"
+            url.endsWith("jpg") -> "image/jpeg"
+            else -> throw IllegalArgumentException("Extension not supported")
+        }
+
+class SponsorDto(
+        val login: String,
+        val company: String,
+        val logoUrl: String,
+        val logoType: String,
+        val logoWebpUrl: String? = null
+)
+
+class EventSponsoringDto(
+        val level: SponsorshipLevel,
+        val sponsor: UserDto,
+        val subscriptionDate: LocalDate = LocalDate.now()
+)
+
+fun EventSponsoring.toDto(language: Language, markdownConverter: MarkdownConverter) =
+        EventSponsoringDto(level, sponsor.toDto(language, markdownConverter), subscriptionDate)
