+++ src/main/kotlin/mixit/model/Event.kt
@@ -37,42 +37,3 @@ enum class SponsorshipLevel {
     NONE
 }
 
-class SponsorDto(
-    val login: String,
-    val company: String,
-    val logoUrl: String,
-    val logoType: String,
-    val logoWebpUrl: String? = null
-)
-
-fun EventSponsoring.toDto() = SponsorDto(
-    this.sponsor.login,
-    this.sponsor.company!!,
-    this.sponsor.logoUrl!!,
-    logoType(this.sponsor.logoUrl),
-    logoWebpUrl(this.sponsor.logoUrl)
-)
-
-private fun logoWebpUrl(url: String) =
-        when {
-            url.endsWith("png") -> url.replace("png", "webp")
-            url.endsWith("jpg") -> url.replace("jpg", "webp")
-            else -> null
-        }
-
-private fun logoType(url: String) =
-        when {
-            url.endsWith("svg") -> "image/svg+xml"
-            url.endsWith("png") -> "image/png"
-            url.endsWith("jpg") -> "image/jpeg"
-            else -> throw IllegalArgumentException("Extension not supported")
-        }
-
-class EventSponsoringDto(
-    val level: SponsorshipLevel,
-    val sponsor: UserDto,
-    val subscriptionDate: LocalDate = LocalDate.now()
-)
-
-fun EventSponsoring.toDto(language: Language, markdownConverter: MarkdownConverter) =
-        EventSponsoringDto(level, sponsor.toDto(language, markdownConverter), subscriptionDate)
