+++ src/main/kotlin/mixit/model/Talk.kt
@@ -56,27 +56,3 @@ enum class Room(name: String, capacity: Int) {
         }
     }
 }
-
-class TalkDto(
-    val id: String?,
-    val slug: String,
-    val format: SessionFormat,
-    val event: String,
-    val title: String,
-    val summary: String,
-    val speakers: List<User>,
-    val language: Language,
-    val addedAt: LocalDateTime,
-    val description: String?,
-    val video: String?,
-    val room: Room?,
-    val start: LocalDateTime?,
-    val end: LocalDateTime?
-)
-
-fun Talk.toDto(markdownConverter: MarkdownConverter) = TalkDto(
-        id, slug, format, event, title,
-        markdownConverter.toHTML(summary), speakers, language, addedAt,
-        markdownConverter.toHTML(description),
-        video, room, start, end
-)
